+++ MAI_OOP_01/Square.h
@@ -8,14 +8,14 @@ class Square : public Figure {
 public:
 	Square();
 	Square(std::istream &is);
-	Square(size_t i);
+	Square(long int i);
 
 	double SquareF() override;
 	void   Print() override;
 
 	virtual ~Square();
 private:
-	size_t side_a;
+	long int side_a;
 };
 
 #endif	/* SQUARE_H */
