+++ MAI_OOP_01/Trapeze.h
@@ -8,16 +8,16 @@ class Trapeze : public Figure {
 public:
 	Trapeze();
 	Trapeze(std::istream &is);
-	Trapeze(size_t i, size_t j, size_t k);
+	Trapeze(long int i, long int j, long int k);
 
 	double SquareF() override;
 	void   Print() override;
 
 	virtual ~Trapeze();
 private:
-	size_t side_a;
-	size_t side_b;
-	size_t height;
+	long int side_a;
+	long int side_b;
+	long int height;
 };
 
 #endif	/* TRAPEZE_H */
